# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: SkillPath API
  version: v1
  description: SkillPath identity, goal, graph, assessment, progress, and review contracts through Phase 4.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/knowledge/nodes/{nodeId}:
    get:
      operationId: getPublishedKnowledgeNode
      parameters:
        - $ref: '#/components/parameters/NodeId'
        - $ref: '#/components/parameters/AcceptLanguage'
      responses:
        '200':
          description: A node in a published graph.
          headers:
            Content-Language:
              $ref: '#/components/headers/ContentLanguage'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeNodeResponse'
        '404':
          $ref: '#/components/responses/Problem'
  /api/v1/knowledge/nodes/{nodeId}/prerequisites:
    get:
      operationId: getPublishedPrerequisites
      parameters:
        - $ref: '#/components/parameters/NodeId'
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/Transitive'
        - $ref: '#/components/parameters/GraphLimit'
      responses:
        '200':
          description: Stable direct or transitive prerequisite nodes.
          headers:
            Content-Language:
              $ref: '#/components/headers/ContentLanguage'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KnowledgeNodeResponse'
  /api/v1/knowledge/nodes/{nodeId}/dependents:
    get:
      operationId: getPublishedDependents
      parameters:
        - $ref: '#/components/parameters/NodeId'
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/Transitive'
        - $ref: '#/components/parameters/GraphLimit'
      responses:
        '200':
          description: Stable direct or transitive dependent nodes.
          headers:
            Content-Language:
              $ref: '#/components/headers/ContentLanguage'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KnowledgeNodeResponse'
...
  /api/v1/knowledge/me:
    get:
      operationId: listMyKnowledgeState
      security: [{ cookieSession: [] }]
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200':
          description: Learner-owned effective knowledge-state page.
          content: { application/json: { schema: { $ref: '#/components/schemas/KnowledgeStatePageResponse' } } }
        '401': { $ref: '#/components/responses/Problem' }
  /api/v1/knowledge/me/{nodeId}:
    get:
      operationId: getMyKnowledgeState
      security: [{ cookieSession: [] }]
      parameters:
        - { name: nodeId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Learner-owned knowledge state for one concept.
          content: { application/json: { schema: { $ref: '#/components/schemas/KnowledgeStateResponse' } } }
        '404': { $ref: '#/components/responses/Problem' }
  /api/v1/knowledge/me/{nodeId}/evidence:
    get:
      operationId: listMyKnowledgeEvidence
      security: [{ cookieSession: [] }]
      parameters:
        - { name: nodeId, in: path, required: true, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200':
          description: Append-only evidence provenance page.
          content: { application/json: { schema: { $ref: '#/components/schemas/KnowledgeEvidencePageResponse' } } }
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: SKILLPATH_SESSION
  parameters:
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: Preferred learner presentation language. Unsupported values fall back to English.
      schema:
        type: string
        enum: [vi-VN, en]
...
    NodeId:
      name: nodeId
      in: path
      required: true
      schema: { type: string }
...
    Transitive:
      name: transitive
      in: query
      schema: { type: boolean, default: false }
    GraphLimit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 100 }
  headers:
    ContentLanguage:
      description: Resolved language used for localized learner content.
      schema:
        type: string
        enum: [vi-VN, en]
  responses:
    Problem:
      description: Request failed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiProblem'
  schemas:
...
    KnowledgeNodeResponse:
      type: object
      required: [id, graphVersionId, slug, name, description, category, difficulty, estimatedMinutes, status]
      properties:
        id: { type: string }
        graphVersionId: { type: string }
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        category: { type: string }
        difficulty: { type: integer, minimum: 1, maximum: 5 }
        estimatedMinutes: { type: integer, minimum: 1 }
        status: { type: string, enum: [DRAFT, ACTIVE, DEPRECATED, ARCHIVED] }
...
    FieldError:
      type: object
      required: [field, code]
      properties:
        field: { type: string }
        code: { type: string }
    ApiProblem:
      type: object
      required: [type, title, status, code, detail, instance, correlationId, fieldErrors]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        instance: { type: string }
        correlationId: { type: string }
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    KnowledgeStateResponse:
      type: object
      required: [knowledgeNodeId, knowledgeNodeSlug, knowledgeNodeName, graphVersionId, recognition, understanding, recall, application, storedMastery, effectiveMastery, confidence, evidenceCount, status, policyVersion]
      properties:
        knowledgeNodeId: { type: string }
        knowledgeNodeSlug: { type: string }
        knowledgeNodeName: { type: string }
        graphVersionId: { type: string }
        recognition: { type: number, minimum: 0, maximum: 1 }
        understanding: { type: number, minimum: 0, maximum: 1 }
        recall: { type: number, minimum: 0, maximum: 1 }
        application: { type: number, minimum: 0, maximum: 1 }
        storedMastery: { type: number, minimum: 0, maximum: 1 }
        effectiveMastery: { type: number, minimum: 0, maximum: 1 }
        confidence: { type: number, minimum: 0, maximum: 1 }
        evidenceCount: { type: integer, minimum: 0 }
        lastEvidenceAt: { type: string, format: date-time }
        nextReviewAt: { type: string, format: date-time }
        status: { type: string, enum: [UNKNOWN, LEARNING, PROVISIONAL, MASTERED, REVIEW_DUE] }
        policyVersion: { type: string }
    KnowledgeStatePageResponse:
      type: object
      required: [items, hasMore]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/KnowledgeStateResponse' } }
        hasMore: { type: boolean }
        nextCursor: { type: string }
    KnowledgeEvidenceResponse:
      type: object
      required: [evidenceId, sourceType, sourceId, dimension, score, reliability, policyVersion, observedAt]
      properties:
        evidenceId: { type: string }
        sourceType: { type: string }
        sourceId: { type: string }
        dimension: { type: string, enum: [RECOGNITION, UNDERSTANDING, RECALL, APPLICATION] }
        score: { type: number, minimum: 0, maximum: 1 }
        reliability: { type: number, minimum: 0, maximum: 1 }
        policyVersion: { type: string }
        observedAt: { type: string, format: date-time }
    KnowledgeEvidencePageResponse:
      type: object
      required: [items, hasMore]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/KnowledgeEvidenceResponse' } }
        hasMore: { type: boolean }
        nextCursor: { type: string }
...